Moved fitting functions to a separate file
