+++ pyxrf/xanes_maps/xanes_maps_api.py
@@ -12,7 +12,8 @@ from ..model.load_data_from_db import make_hdf
 from ..model.command_tools import pyxrf_batch
 from ..model.fileio import read_hdf_APS
 from ..model.utils import (grid_interpolate, normalize_data_by_scaler, convert_time_to_nexus_string,
-                           check_if_eline_is_activated, check_eline_name, fit_spectrum, rfactor_compute)
+                           check_if_eline_is_activated, check_eline_name)
+from ..core.fitting import fit_spectrum, rfactor_compute
 
 import logging
 logger = logging.getLogger()
